move mongoose models into dedicated mongoose folder
